Deprecate repository category (initial)
All public API and visuals related to repository categories are removed.
This will pave the way to restructure of how repository information is
structured, persisted and presented to the user.

NB: some changes to the Dashboard are artificial to make the code compile.
A number of elements have been deleted and will be restored in subsequent
PRs dealing with the UI overhaul.
Also there are unused `SerializeRepositories` and `DeserializeRepositories`
methods left behind, and they will be dealt with in subsequent PRs as well
to provide a migration path from the old settings structure to the new.