Fix restoring expanded state issues when TreeNode.Text changes (e.g. submodule nodes show branch and ahead/behind text)
Added TreeViewExtensions.GetFullNamePath, which returns a full path to the node preferring to use the node's Name over it's Text field. GetExpandedNodeState now uses this function instead of TreeNode.FullPath. This allows us to vary the node's Text while keeping a stable Name. Along with this change, SubmoduleNode now returns the name of the submodule as the node's Name, which fixes issues of its nodes not being restored as expected when performing actions that would change the Text (e.g. changing branch, etc.).